Manual

Model Specification

Recursion

Links

Links are reparametrizations utilized to ensure certain parameter is within its original domain, i.e. in a distribution one would like to ensure that the time varying parameter $f \in \mathbb{R}^+$. The way to do this is to model $\tilde{f} = \ln{f}$. More generally one can stablish that $\tilde{f} = h(f)$. We refer to this procedure as linking. When the parameter is linked the GAS recursion happens in the domain of $\tilde{f}$ and then one can recover the orginal parameter by $f = \left(h\right)^-1(\tilde f)$. We refer to this procedure as unlinking. The new GAS recursion becomes.

\[\begin{equation*}\left\{\begin{array}{ccl} f_{t} &=& h^{-1}(\tilde f_t), \\ \tilde f_{t+1} &=& \omega + \sum_{i=1}^p A_{i}\tilde s_{t-i+1} + \sum_{j=1}^q B_{j}\tilde f_{t-j+1} \end{array} \right. \end{equation*}\]

Notice that the change in parametrization changes the dynamics of the model. The GAS(1,1) for a Normal distribution with inverse scaling $d = 1$ is equivalent to the GARCH(1, 1) model, but only on the original parameter, if you work with a different parametrization the model is no longer equivalent.

Types of links

The abstract type Link subsumes any type of link that can be expressed.

ScoreDrivenModels.IdentityLink — Type

IdentityLink <: Link

Define the map $\tilde{f} = f$ where $f \in \mathbb{R}$ and $\tilde{f} \in \mathbb{R}$

ScoreDrivenModels.LogLink — Type

LogLink <: Link

Define the map $\tilde{f} = \ln(f - a)$ where $f \in [a, \infty), a \in \mathbb{R}$ and $\tilde{f} \in \mathbb{R}$

ScoreDrivenModels.LogitLink — Type

LogitLink <: Link

Define the map $\tilde{f} = \-ln(\frac{b - a}{f + a} - 1)$ where $f \in [a, b], a, b \in \mathbb{R}$ and $\tilde{f} \in \mathbb{R}$

Link functions

ScoreDrivenModels.link — Function

link(args...)

The link function is a map that brings a parameter $f$ in a subspace $\mathcal{F} \subset \mathbb{R}$ to $\mathbb{R}$.

ScoreDrivenModels.unlink — Function

unlink(args...)

The unlink function is the inverse map of link. It brings $\tilde{f}$ in $\mathbb{R}$ to the subspace $\mathcal{F} \subset \mathbb{R}$.

ScoreDrivenModels.jacobian_link — Function

jacobian_link(args...)

Evaluates the derivative of the link with respect to the parameter $f$.