Manual

Links

Links are reparametrizations utilized in the estimation. For example, if we want to estimate a parameter $f$ which is by definition strictly positive, then an obvious way to estimate $f$ via numerical optimization is to model $\tilde{f} = \ln{f}$. We refer to this procedure as linking. After obtaining the optimal value of $\tilde{f}$, we can then unlink it to obtain $f$ by computing $f = e^{\tilde{f}}$.

Types of links

The abstract type Link subsumes any type of link that can be expressed.

ScoreDrivenModels.IdentityLink — Type

IdentityLink <: Link

Define the map $\tilde{f} = f$ where $f \in \mathbb{R}$ and $\tilde{f} \in \mathbb{R}$

ScoreDrivenModels.LogLink — Type

LogLink <: Link

Define the map $\tilde{f} = \ln(f - a)$ where $f \in [a, \infty), a \in \mathbb{R}$ and $\tilde{f} \in \mathbb{R}$

Link functions

ScoreDrivenModels.link — Function

link(args...)

The link function is a map that brings a parameter $f$ in a subspace $\mathcal{F} \subset \mathbb{R}$ to $\mathbb{R}$.

ScoreDrivenModels.unlink — Function

unlink(args...)

The unlink function is the inverse map of link. It brings $\tilde{f}$ in $\mathbb{R}$ to the subspace $\mathcal{F} \subset \mathbb{R}$.

ScoreDrivenModels.jacobian_link — Function

jacobian_link(args...)

Evaluates the derivative of the link with respect to the parameter $f$.